refactor(data): merge the loot parser trio into one LootTableParser
LootTableParser, PoolParser, and EntryParser were three classes wired
together through mutual lateinit vars set in an init block — mandatory
post-construction surgery for what is just mutual recursion in the loot
format (tables contain pools contain entries which reference tables).
Fold pools and entries into LootTableParser(path, version); the
recursion becomes plain method calls and the lateinit wiring disappears.

parseReferencedTable now JSON-parses + parses the referenced table
directly instead of round-tripping through ExtractLootTables.parseFile;
referenced tables therefore skip the loot-type whitelist, which is fine
because only their produced/required items are used, never their type.

Behavior unchanged; the three test files keep their per-level coverage
and only swap construction.
